+++ khoros/core.py
@@ -6,7 +6,7 @@
 :Example:           ``khoros = Khoros(helper='helper.yml')``
 :Created By:        Jeff Shurtliff
 :Last Modified:     Jeff Shurtliff
-:Modified Date:     07 Sep 2023
+:Modified Date:     11 Sep 2023
 """
 
 import sys
@@ -1741,11 +1741,11 @@ class Khoros(object):
             """
             return structures_module.boards.get_message_count(self.khoros_object, board_id)
 
-        def get_all_messages(self, board_id, fields=None, where_filter=None):
+        def get_all_messages(self, board_id, fields=None, where_filter=None, descending=True):
             """This function retrieves data for all messages within a given board.
 
             .. versionchanged:: 5.4.0
-               Introduced the ``where_filter`` parameter to optionally further filter the LiQL query.
+               Introduced the ``where_filter`` and ``descending`` parameters to optionally filter the LiQL query.
 
             .. versionadded:: 5.3.0
 
@@ -1755,12 +1755,15 @@ class Khoros(object):
             :type fields: str, tuple, list, set, None
             :param where_filter: One or more optional WHERE filters to include in the LiQL query
             :type where_filter: str, tuple, list, set, None
+            :param descending: Determines if the data should be returned in descending order (``True`` by default)
+            :type descending: bool
             :returns: A list containing a dictionary of data for each message within the board
             :raises: :py:exc:`khoros.errors.exceptions.GETRequestError`
             """
-            return structures_module.boards.get_all_messages(self.khoros_object, board_id, fields, where_filter)
+            return structures_module.boards.get_all_messages(self.khoros_object, board_id, fields,
+                                                             where_filter, descending)
 
-        def get_all_topic_messages(self, board_id, fields=None):
+        def get_all_topic_messages(self, board_id, fields=None, descending=True):
             """This function retrieves data for all topic messages (i.e. zero-depth messages) within a given board.
 
             .. versionadded:: 5.4.0
@@ -1769,10 +1772,12 @@ class Khoros(object):
             :type board_id: str
             :param fields: Specific fields to query if not all fields are needed (comma-separated string or iterable)
             :type fields: str, tuple, list, set, None
+            :param descending: Determines if the data should be returned in descending order (``True`` by default)
+            :type descending: bool
             :returns: A list containing a dictionary of data for each topic message within the board
             :raises: :py:exc:`khoros.errors.exceptions.GETRequestError`
             """
-            return structures_module.boards.get_all_topic_messages(self.khoros_object, board_id, fields)
+            return structures_module.boards.get_all_topic_messages(self.khoros_object, board_id, fields, descending)
 
     class BulkData(object):
         """This class includes methods for interacting with the Bulk Data API.
@@ -3391,7 +3396,7 @@ class Khoros(object):
             """
             return objects_module.messages.format_user_mention(self.khoros_object, user_info, user_id, login)
 
-        def get_all_messages(self, board_id, fields=None, where_filter=None):
+        def get_all_messages(self, board_id, fields=None, where_filter=None, descending=True):
             """This function retrieves data for all messages within a given board.
 
             .. versionadded:: 5.4.0
@@ -3402,12 +3407,15 @@ class Khoros(object):
             :type fields: str, tuple, list, set, None
             :param where_filter: One or more optional WHERE filters to include in the LiQL query
             :type where_filter: str, tuple, list, set, None
+            :param descending: Determines if the data should be returned in descending order (``True`` by default)
+            :type descending: bool
             :returns: A list containing a dictionary of data for each message within the board
             :raises: :py:exc:`khoros.errors.exceptions.GETRequestError`
             """
-            return objects_module.messages.get_all_messages(self.khoros_object, board_id, fields, where_filter)
+            return objects_module.messages.get_all_messages(self.khoros_object, board_id, fields,
+                                                            where_filter, descending)
 
-        def get_all_topic_messages(self, board_id, fields=None):
+        def get_all_topic_messages(self, board_id, fields=None, descending=True):
             """This function retrieves data for all topic messages (i.e. zero-depth messages) within a given board.
 
             .. versionadded:: 5.4.0
@@ -3416,10 +3424,12 @@ class Khoros(object):
             :type board_id: str
             :param fields: Specific fields to query if not all fields are needed (comma-separated string or iterable)
             :type fields: str, tuple, list, set, None
+            :param descending: Determines if the data should be returned in descending order (``True`` by default)
+            :type descending: bool
             :returns: A list containing a dictionary of data for each topic message within the board
             :raises: :py:exc:`khoros.errors.exceptions.GETRequestError`
             """
-            return objects_module.messages.get_all_topic_messages(self.khoros_object, board_id, fields)
+            return objects_module.messages.get_all_topic_messages(self.khoros_object, board_id, fields, descending)
 
 
     class Node(object):